refactor(exporter): promote simulate + source + inheritance graph to pkg/config (PR-8)
v2.8.0 PR-4 made pkg/config the canonical location for resolver +
parse + inheritance helpers. PR-8 closes the library shape: promote
the remaining hierarchy + simulate primitives so cmd/da-guard,
tenant-api, and any future component can import them directly without
going through `package main` or HTTP.

## File-level changes

CREATED:
  - pkg/config/inheritance_graph.go (123 LoC) — InheritanceGraph type
    + NewInheritanceGraph + AddTenant + TenantsAffectedBy methods +
    CollectDefaultsChain (capital export). app's scanDirHierarchical
    + pkg/config's ScanFromConfigSource both call CollectDefaultsChain.

RENAMED (git rename detected, 100% similarity except package decl):
  - app/config_source.go → pkg/config/source.go (206 LoC)
    ConfigSource interface + InMemoryConfigSource + ScanFromConfigSource
    (capitalized — exported for test callers). Internal call to
    scanFromConfigSource updated to ScanFromConfigSource.
  - app/config_simulate.go → pkg/config/simulate.go (149 LoC)
    SimulateRequest / SimulateResponse / ErrSimulateTenantNotFound /
    SimRoot / SimulateEffective. Internal calls switched to capital
    exports: ComputeEffectiveConfig / ComputeMergedHash /
    ComputeSourceHash (already exported by PR-4).

UPDATED:
  - app/config_hierarchy.go — removed InheritanceGraph type +
    NewInheritanceGraph + AddTenant + TenantsAffectedBy +
    collectDefaultsChain + sortStrings. Replaced internal calls with
    config.CollectDefaultsChain + sort.Strings (stdlib). Net -118 LoC.
  - app/config_types.go — added InheritanceGraph type alias +
    NewInheritanceGraph var alias so existing field types
    (`*InheritanceGraph`) and constructor calls in config.go /
    config_debounce.go / tests compile unchanged.
  - app/handler_simulate.go — switched to config.SimulateRequest +
    config.SimulateEffective + config.ErrSimulateTenantNotFound.
    Transport stays in app/, semantic logic now in pkg/config.
  - app/config_simulate_test.go — call sites updated (config.X
    namespace) so tests still exercise the moved code.
  - app/config_hierarchy_test.go — sortStrings → sort.Strings (stdlib).

## Library shape achieved

After PR-8, pkg/config is self-contained for:
  - Resolution (PR-4): ApplyProfiles, ResolveAt, ResolveStateFiltersAt,
    etc.
  - Parse helpers (PR-4): ParseHHMM, ParseMetricKey, ParseKeyWithLabels,
    MatchTimeWindow, ClampDuration, FormatDuration, IsDisabled, etc.
  - Inheritance helpers (PR-4): DeepMerge, DeepCopyMap,
    NormalizeYAMLToJSON, ExtractDefaultsBlock, ExtractTenantRaw,
    CanonicalJSON, ComputeEffectiveConfig, ComputeMergedHash,
    ComputeSourceHash.
  - Hierarchy graph (PR-8): InheritanceGraph + CollectDefaultsChain.
  - Config source (PR-8): ConfigSource interface + InMemoryConfigSource
    + ScanFromConfigSource.
  - Simulate primitive (PR-8): SimulateEffective + types.

cmd/da-guard's `da-tools guard defaults-impact` predict-flow could now
call config.ScanFromConfigSource + config.SimulateEffective directly
instead of shell-out — left as a follow-up since it's a behavior
change, not a refactor.

app/ keeps:
  - ConfigManager (orchestrator) — disk I/O, watchloop, debounce
  - scanDirHierarchical — disk-walking scanner with mtime fast-path
    (uses fileStat which is exporter-specific)
  - DuplicateTenantError — scanner error type
  - Handlers (HTTP transport)
  - flat_scanner.go — flat-mode legacy support

## Verification

- go build -buildvcs=false ./... clean
- go vet ./... clean
- Linux dev container worktree path:
  go test -race -count=1 ./... — all 9 packages pass (5.7s)
- Hot-path stress -race -count=10 (TestSlowWriteTornStateStress +
  TestDebounce + TestDiffAndReload + TestComputeMergedHash +
  TestGoldenParity + TestSimulate*) — 22.4s, 0 races, 0 flakes
- Apples-to-apples bench (Windows host, parallel git worktree, count=2):
    DiffAndReload_Hierarchical_1000_NoChange: 398ms → 399ms (+0.2%)
    DiffAndReload_Hierarchical_2000_NoChange: 749ms → 738ms (-1.6%)
  Within noise envelope.
- Golden parity test (TestGoldenParity*) confirms 16-char merged_hash
  output is byte-equal pre/post promote — same as PR-4 discipline.

## Net delta

- app/ side: -497 LoC (config_source.go -205, config_simulate.go -149,
  config_hierarchy.go -118 inheritance helpers, +misc -25)
- pkg/config/ side: +478 LoC (inheritance_graph.go +123, simulate.go
  +149, source.go +206)
- Net: -19 LoC, but the value is library shape — pkg/config now usable
  as a standalone library for any cross-component consumer.
